# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/public/api/add-customer": {
            "post": {
                "tags": [
                    "Public Customers"
                ],
                "summary": "Add new Customer with a Quick Lead",
                "operationId": "238b2b3b3b6e20a76c363681f910bde1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "": {
                                        "$ref": "#/components/schemas/CustomerQLSchema"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Vincent Mark",
                                    "email": "vmbcarabbacan@gmail.com",
                                    "country_code": "+971",
                                    "phone_number": "566368779",
                                    "other_contact_info": null,
                                    "insurance_type": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer added/updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Vincent Mark"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "vmbcarabbacan@gmail.com"
                                                },
                                                "country_code": {
                                                    "type": "string",
                                                    "example": "+971"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "566368779"
                                                },
                                                "other_contact_info": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T16:39:54.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T16:39:54.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": 18
                                                },
                                                "created_by": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "complete_phone_number": {
                                                    "type": "string",
                                                    "example": "971566368779"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "maxItems": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CustomerQLSchema": {
                "properties": {
                    "": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "country_code": {
                                "type": "string"
                            },
                            "phone_number": {
                                "type": "string"
                            },
                            "other_contact_info": {
                                "type": "string"
                            },
                            "insurance_type": {
                                "type": "number"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}